refactor: remove move identity-migration module in migration package
