+++ migration/identity-migration/src/test/java/io/camunda/migration/identity/AuthorizationMigrationHandlerTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.identity.migration;
+package io.camunda.migration.identity;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -14,7 +14,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import io.camunda.identity.migration.dto.UserResourceAuthorization;
+import io.camunda.migration.identity.dto.UserResourceAuthorization;
 import io.camunda.service.AuthorizationServices;
 import java.util.Collection;
 import java.util.List;
