+++ parent/pom.xml
@@ -1077,12 +1077,6 @@
 
       <dependency>
         <groupId>io.camunda</groupId>
-        <artifactId>identity-migration</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-
-      <dependency>
-        <groupId>io.camunda</groupId>
         <artifactId>identity-spring-boot-starter</artifactId>
         <version>${version.identity}</version>
       </dependency>
@@ -2265,7 +2259,7 @@
 
       <dependency>
         <groupId>io.camunda.migration</groupId>
-        <artifactId>migration-api</artifactId>
+        <artifactId>identity-migration</artifactId>
         <version>${project.version}</version>
       </dependency>
